rafactor: add exclude decorator for user entity
