+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { ArtistModule } from './artist/artist.module';
-import { AlbumModule } from './album/album.module';
-import { TrackModule } from './track/track.module';
-import { UserModule } from './user/user.module';
-import { FavoritesModule } from './favorites/favorites.module';
+import { ArtistModule } from './entities/artist/artist.module';
+import { AlbumModule } from './entities/album/album.module';
+import { TrackModule } from './entities/track/track.module';
+import { UserModule } from './entities/user/user.module';
+import { FavoritesModule } from './entities/favorites/favorites.module';
 
 @Module({
   imports: [
