+++ src/entities/album/album.interface.ts
