+++ src/entities/artist/artist.controller.ts
