+++ src/entities/artist/artist.dto.ts
