+++ src/entities/artist/artist.interface.ts
