+++ src/entities/artist/artist.module.ts
@@ -2,9 +2,9 @@ import { forwardRef, Module } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
 import { ValidateArtistPipe } from './validate-artist.pipe';
-import { TrackModule } from 'src/track/track.module';
-import { AlbumModule } from 'src/album/album.module';
-import { FavoritesModule } from 'src/favorites/favorites.module';
+import { TrackModule } from 'src/entities/track/track.module';
+import { AlbumModule } from 'src/entities/album/album.module';
+import { FavoritesModule } from 'src/entities/favorites/favorites.module';
 
 @Module({
   imports: [
