+++ src/entities/artist/artist.service.ts
@@ -2,9 +2,9 @@ import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
 import { Artist } from './artist.interface';
 import { createArtistDto, updateArtistDto } from './artist.dto';
-import { TrackService } from 'src/track/track.service';
-import { AlbumService } from 'src/album/album.service';
-import { FavoritesService } from 'src/favorites/favorites.service';
+import { TrackService } from 'src/entities/track/track.service';
+import { AlbumService } from 'src/entities/album/album.service';
+import { FavoritesService } from 'src/entities/favorites/favorites.service';
 
 @Injectable()
 export class ArtistService {
