+++ src/entities/artist/validate-artist.pipe.ts
