+++ src/entities/favorites/favorites.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { TrackModule } from 'src/track/track.module';
-import { AlbumModule } from 'src/album/album.module';
-import { ArtistModule } from 'src/artist/artist.module';
+import { TrackModule } from 'src/entities/track/track.module';
+import { AlbumModule } from 'src/entities/album/album.module';
+import { ArtistModule } from 'src/entities/artist/artist.module';
 import { FavoritesService } from './favorites.service';
 import { FavoritesController } from './favorites.controller';
 import {
