+++ src/entities/track/track.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Module } from '@nestjs/common';
 import { TrackService } from './track.service';
 import { ValidateTrackPipe } from './validate-track.pipe';
 import { TrackController } from './track.controller';
-import { FavoritesModule } from 'src/favorites/favorites.module';
+import { FavoritesModule } from 'src/entities/favorites/favorites.module';
 
 @Module({
   imports: [forwardRef(() => FavoritesModule)],
