+++ src/entities/user/user.controller.ts
@@ -12,9 +12,11 @@ import {
   HttpStatus,
   UsePipes,
   ValidationPipe,
+  UseInterceptors,
+  ClassSerializerInterceptor,
 } from '@nestjs/common';
 import { UserService } from './user.service';
-import { CreateUserDto, UpdatePasswordDto } from './user.dto';
+import { CreateUserDto, UpdatePasswordDto, UserEntity } from './user.dto';
 import { ERRORS } from 'src/consts/ERRORS';
 import { ValidateUserPipe } from './validate-user.pipe';
 
@@ -22,44 +24,39 @@ import { ValidateUserPipe } from './validate-user.pipe';
 export class UserController {
   constructor(private service: UserService) {}
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Get()
   async findAll() {
     const users = await this.service.getAll();
     const response = users.map((user) => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const { password, ...newUser } = user;
-      return newUser;
+      return new UserEntity(user);
     });
     return response;
   }
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Get(':id')
   async findOne(@Param('id', ParseUUIDPipe) id: string) {
     const user = await this.service.get(id);
     if (!user) {
       throw new HttpException(ERRORS.NOT_FOUND('User'), HttpStatus.NOT_FOUND);
     }
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { password, ...newUser } = user;
-    return newUser;
+    return new UserEntity(user);
   }
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Post()
   async create(@Body() createDto: CreateUserDto) {
     const user = await this.service.create(createDto);
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { password, ...newUser } = user;
-    return newUser;
+    return new UserEntity(user);
   }
 
+  @UseInterceptors(ClassSerializerInterceptor)
   @Put(':id')
   @UsePipes(ValidateUserPipe, ValidationPipe)
   async update(@Body() updateDto: UpdatePasswordDto) {
     const updatedUser = await this.service.update(updateDto);
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { password, ...newUser } = updatedUser;
-    return newUser;
+    return new UserEntity(updatedUser);
   }
 
   @Delete(':id')
