+++ src/entities/user/user.interface.ts
