--- src/favorites/favorites.interface.ts
@@ -1,9 +1,0 @@
-import { Album } from 'src/album/album.interface';
-import { Artist } from 'src/artist/artist.interface';
-import { Track } from 'src/track/track.interface';
-
-export interface Favorites {
-  artists: Artist[]; // favorite artists ids
-  albums: Album[]; // favorite albums ids
-  tracks: Track[]; // favorite tracks ids
-}
