--- src/user/user.dto.ts
@@ -1,20 +1,0 @@
-import { IsNotEmpty, IsUUID } from 'class-validator';
-
-export class CreateUserDto {
-  @IsNotEmpty()
-  login: string;
-
-  @IsNotEmpty()
-  password: string;
-}
-
-export class UpdatePasswordDto {
-  @IsUUID()
-  id: string;
-
-  @IsNotEmpty()
-  oldPassword: string; // previous password
-
-  @IsNotEmpty()
-  newPassword: string; // new password
-}
